requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
uvicorn==0.29.0
whitenoise==6.6.0
